+++ tests/common.rs
@@ -8,6 +8,20 @@ pub fn test_public_key_path() -> std::path::PathBuf {
     cargo_manifest_dir().join("test_assets/public_key.asc")
 }
 
+pub fn test_protected_private_key_path() -> std::path::PathBuf {
+    cargo_manifest_dir()
+        .join("test_assets/fixture_packages/signing_keys/secret_rsa4096_protected.asc")
+}
+
+pub fn test_protected_public_key_path() -> std::path::PathBuf {
+    cargo_manifest_dir()
+        .join("test_assets/fixture_packages/signing_keys/public_rsa4096_protected.asc")
+}
+
+pub fn test_protected_private_key_passphrase() -> String {
+    "thisisN0Tasecuredpassphrase".to_owned()
+}
+
 pub fn rpm_389_ds_file_path() -> std::path::PathBuf {
     cargo_manifest_dir().join("test_assets/389-ds-base-devel-1.3.8.4-15.el7.x86_64.rpm")
 }
@@ -52,6 +66,14 @@ pub fn rsa_public_key() -> Vec<u8> {
     let public_key =
         cargo_manifest_dir().join("test_assets/fixture_packages/signing_keys/public_rsa4096.asc");
     std::fs::read(public_key).unwrap()
+}
+
+pub fn load_protected_rsa_keys() -> (Vec<u8>, Vec<u8>) {
+    let signing_key =
+        include_bytes!("../test_assets/fixture_packages/signing_keys/secret_rsa4096_protected.asc");
+    let verification_key =
+        include_bytes!("../test_assets/fixture_packages/signing_keys/public_rsa4096_protected.asc");
+    (signing_key.to_vec(), verification_key.to_vec())
 }
 
 pub fn eddsa_private_key() -> Vec<u8> {